fix blog
